refactor: 타이머 캐러셀 컴포넌트 정리 (#308)
